Adapt platform controller in the platforms middleware with configs in project skeleton
