--- bindings/nodejs/lib/middleware/platforms/config/general.json
@@ -1,3 +1,0 @@
-{
-    "baseUrl":"https://mydomain.com"
-}
