+++ bindings/nodejs/lib/middleware/platforms/controllers/kik.js
@@ -41,12 +41,6 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 
 let Bot  = require('@kikinteractive/kik');
-var kikConfig = require('../config/kik/default.json');
-
-// Configure the bot API endpoint, details for your bot
-let bot = new Bot(kikConfig);
-
-bot.updateBotConfiguration();
 
 var replyHandler = function(message, reply) {
     message.reply(reply);
@@ -64,7 +58,12 @@ var messageHandler = function(chatbotClient, message) {
 };
 
 //Weird and all but I guess this route is enforced by the lib
-module.exports.attach = function (chatbotClient, app) {
+module.exports.attach = function (chatbotClient, kikConfig, app) {
+    // Configure the bot API endpoint, details for your bot
+    let bot = new Bot(kikConfig);
+
+    bot.updateBotConfiguration();
+
     bot.onTextMessage((message) => {
         messageHandler(chatbotClient, message);
     });
