+++ bindings/nodejs/project-skeletons/chatbot/app.js
@@ -40,6 +40,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 var chatbot =  require('./chatbot');
 var openintent = require('open-intent');
+var config = require('./config');
 
 var REST_PORT = 8080;
 
@@ -47,6 +48,6 @@ var middlewares = [];
 
 middlewares.push(openintent.middleware.irc());
 middlewares.push(openintent.middleware.rest(REST_PORT));
-//middlewares.push(openintent.middleware.platforms());
+middlewares.push(openintent.middleware.platforms(config));
 
 chatbot(middlewares);
