+++ bindings/nodejs/project-skeletons/chatbot/config/general.json
@@ -1,0 +1,4 @@
+{
+    "baseUrl":"https://mydomain.com",
+    "port": 5000
+}
