+++ bindings/nodejs/project-skeletons/chatbot/config/selection.json
