+++ bindings/nodejs/project-skeletons/chatbot/config/skype/default.json
