+++ bindings/nodejs/project-skeletons/chatbot/config/slack/default.json
