+++ package.json
@@ -21,17 +21,22 @@
     "entity"
   ],
   "dependencies": {
+    "@kikinteractive/kik": "^2.0.8",
+    "@slack/client": "^3.5.1",
     "bindings": "1.2.x",
     "body-parser": "^1.15.2",
     "commander": "^2.9.0",
+    "crypto": "0.0.3",
     "express": "^4.14.0",
     "fs-extra": "^0.30.0",
     "lodash": "^4.13.1",
     "nan": "2.3.x",
+    "node-telegram-bot-api": "^0.23.3",
     "nodemon": "^1.10.0",
     "q": "^1.4.1",
     "readline": "^1.3.0",
     "redis": "^2.6.0-2",
+    "request": "^2.73.0",
     "requestify": "^0.1.17",
     "rmdir": "^1.2.0",
     "seq-queue": "0.0.5",
